Simplified command line interface
